Fix .Timeout() with mutexes (broken with #10)
in #10 we added a mutex around .Called(), but this causes an issue when
the mock has .Timeout() set. Currently all callers to the mock are
locked out for this timeout.

As a simple fix, wait to do the sleep until
after we have finished the method and unlocked the mutex.